First Glance
might be best to discuss how the circle and connection are handled in a UML design style.
Tested by creating two users and clicking on the invites/new , filling in the form and then copying the invites id (ex. 0.0.0.0:8000/invites/(uuid)  while logged in to a different account that would not have this id, accepted requested and connection was made. If user 1 adds user to Family and the user 2 adds to their different groups ( for example, they delete family and friend. and User 2 creates Climbing circle and wants to add user 1 to climbing... how do we distinguish in the system that the connection belongs in this circle or that circle?