First commit
Just checking if this is along the lines of what you had in mind.
The tests aren't necessary. I think I would like to see the execution on a browser to really understand what I'm doing but as of now I'm simply creating connection on the back end.

Is there an issue open for creating/enhancing the connection front end so a user can actually see who they submit a request to?